3-test-answer/src/requirements.txt
playwright
bs4
argparse
psycopg2-binary
fastapi
uvicorn